refactor(onboarding): split ConnectGitStep into smaller components
ConnectGitStep was ~900 lines covering three independent concerns. Split into:

- `GitHubConnectPanel` — self-contained GitHub OAuth UI (status header,
  alternative-project copy, connect buttons, integration rows, disconnect
  dialog).
- `CliCheckPanel` + `InstalledBadge` — generic shell for the Git / GitHub
  CLI install panels, deduping the near-identical Box + header + spinner
  +status-badge scaffold.
- `ConnectGitStep` — now ~350 lines, orchestrates the header, benefits
  list, the three panels, and step actions.

No behavior change.

Generated-By: PostHog Code
Task-Id: d15be8ed-c5d9-44cc-8cc7-de90e1e84f2e
